[Refactor] Extract akka http input matcher
