+++ server/akka-http-server/src/main/scala/tapir/server/akkahttp/EndpointToAkkaServer.scala
@@ -6,7 +6,7 @@ import akka.http.scaladsl.model.HttpHeader.ParsingResult
 import akka.http.scaladsl.model.{MediaType => _, _}
 import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.util.{Tuple => AkkaTuple}
-import akka.http.scaladsl.server.{Directive, Directive1, Route}
+import akka.http.scaladsl.server.{Directive, Directive1, RequestContext, Route}
 import tapir._
 import tapir.internal.{ParamsToSeq, SeqToParams}
 import tapir.typelevel.{ParamsAsArgs, ParamsToTuple}
@@ -74,89 +74,10 @@ object EndpointToAkkaServer {
     import akka.http.scaladsl.server.Directives._
     import akka.http.scaladsl.server._
 
-    val methodDirective = e.method match {
-      case Method.GET     => get
-      case Method.HEAD    => head
-      case Method.POST    => post
-      case Method.PUT     => put
-      case Method.DELETE  => delete
-      case Method.OPTIONS => options
-      case Method.PATCH   => patch
-      case m              => method(HttpMethod.custom(m.m))
-    }
+    val methodDirective = methodDirectiveType(e)
 
     // TODO: when parsing a query parameter/header/body/path fragment fails, provide an option to return a nice
     // error to the user (instead of a 404).
-
-    case class MatchResult(values: List[Any], ctx: RequestContext, canRemoveSlash: Boolean) {
-      def prependValue(v: Any): MatchResult = copy(values = v :: values)
-    }
-
-    def doMatch(inputs: Vector[EndpointInput.Single[_]], ctx: RequestContext, canRemoveSlash: Boolean, body: Any): Option[MatchResult] = {
-
-      def handleMapped[II, T](wrapped: EndpointInput[II], f: II => T, inputsTail: Vector[EndpointInput.Single[_]]): Option[MatchResult] = {
-        doMatch(wrapped.asVectorOfSingle, ctx, canRemoveSlash, body).flatMap { result =>
-          doMatch(inputsTail, result.ctx, result.canRemoveSlash, body).map {
-            _.prependValue(f.asInstanceOf[Any => Any].apply(SeqToParams(result.values)))
-          }
-        }
-      }
-
-      inputs match {
-        case Vector() => Some(MatchResult(Nil, ctx, canRemoveSlash))
-        case EndpointInput.PathSegment(ss) +: inputsTail =>
-          ctx.unmatchedPath match {
-            case Uri.Path.Slash(pathTail) if canRemoveSlash =>
-              doMatch(inputs, ctx.withUnmatchedPath(pathTail), canRemoveSlash = false, body)
-            case Uri.Path.Segment(`ss`, pathTail) =>
-              doMatch(inputsTail, ctx.withUnmatchedPath(pathTail), canRemoveSlash = true, body)
-            case _ => None
-          }
-        case EndpointInput.PathCapture(codec, _, _, _) +: inputsTail =>
-          ctx.unmatchedPath match {
-            case Uri.Path.Slash(pathTail) if canRemoveSlash =>
-              doMatch(inputs, ctx.withUnmatchedPath(pathTail), canRemoveSlash = false, body)
-            case Uri.Path.Segment(s, pathTail) =>
-              codec.decode(s) match {
-                case DecodeResult.Value(v) =>
-                  doMatch(inputsTail, ctx.withUnmatchedPath(pathTail), canRemoveSlash = true, body).map {
-                    _.prependValue(v)
-                  }
-                case _ => None
-              }
-            case _ => None
-          }
-        case EndpointInput.Query(name, codec, _, _) +: inputsTail =>
-          codec.decodeOptional(ctx.request.uri.query().get(name)) match {
-            case DecodeResult.Value(v) =>
-              doMatch(inputsTail, ctx, canRemoveSlash = true, body).map {
-                _.prependValue(v)
-              }
-            case _ => None
-          }
-        case EndpointIO.Header(name, codec, _, _) +: inputsTail =>
-          codec.decodeOptional(ctx.request.headers.find(_.is(name.toLowerCase)).map(_.value())) match {
-            case DecodeResult.Value(v) =>
-              doMatch(inputsTail, ctx, canRemoveSlash = true, body).map {
-                _.prependValue(v)
-              }
-            case _ => None
-          }
-        case EndpointIO.Body(codec, _, _) +: inputsTail =>
-          codec.decodeOptional(Some(body)) match {
-            case DecodeResult.Value(v) =>
-              doMatch(inputsTail, ctx, canRemoveSlash = true, body).map {
-                _.prependValue(v)
-              }
-            case _ => None
-          }
-        case EndpointInput.Mapped(wrapped, f, _, _) +: inputsTail =>
-          handleMapped(wrapped, f, inputsTail)
-        case EndpointIO.Mapped(wrapped, f, _, _) +: inputsTail =>
-          handleMapped(wrapped, f, inputsTail)
-      }
-
-    }
 
     val inputDirectives: Directive1[I] = {
       val bodyDirective: Directive1[Any] = bodyType(e.input) match {
@@ -166,7 +87,7 @@ object EndpointToAkkaServer {
       }
       bodyDirective.flatMap { body =>
         extractRequestContext.flatMap { ctx =>
-          doMatch(e.input.asVectorOfSingle, ctx, canRemoveSlash = true, body) match {
+          AkkaHttpInputMatcher.doMatch(e.input.asVectorOfSingle, ctx, body) match {
             case Some(result) =>
               provide(SeqToParams(result.values).asInstanceOf[I]) & mapRequestContext(_ => result.ctx)
             case None => reject
@@ -176,6 +97,19 @@ object EndpointToAkkaServer {
     }
 
     methodDirective & inputDirectives
+  }
+
+  private def methodDirectiveType[O, E, I](e: Endpoint[I, E, O]) = {
+    e.method match {
+      case Method.GET     => get
+      case Method.HEAD    => head
+      case Method.POST    => post
+      case Method.PUT     => put
+      case Method.DELETE  => delete
+      case Method.OPTIONS => options
+      case Method.PATCH   => patch
+      case m              => method(HttpMethod.custom(m.m))
+    }
   }
 
   private def bodyType[I](in: EndpointInput[I]): Option[RawValueType[_]] = {
